refactor(breaker, telemetry): decouple breaker from observability; split metrics
Closes #616 and #611.

#616: internal/breaker/breaker.go used to call telemetry.CircuitBreakerState
and event.Emit directly from inside the state-machine transition. The
breaker is a reusable primitive; it should not know about Prometheus or
the app's event bus.

- Removes the observe imports from internal/breaker.
- Adds CircuitBreaker.SetOnStateChange(func(StateChangeInfo)) for callers
  to install observability hooks.
- StateChangeInfo carries Name, From, To, Failures, Threshold, and
  OpenDuration so the hook has the context it needs.
- New telemetry.NewCircuitBreakerHook(name) returns a callback that
  drives CircuitBreakerState gauge, CircuitBreakerTransitionsTotal
  counter, and the BackendCircuitOpened / BackendCircuitClosed events.
- Wired in store.NewDatabaseBreaker, backend.NewCircuitBreakerBackend,
  and counter.NewRedisCounterBackend.
- Three breaker tests that asserted on the event bus moved to
  telemetry/cb_hook_test.go where the hook lives; equivalent
  hook-invocation tests stay in breaker_test.go.
- internal/breaker now imports nothing under internal/observe.

#611: internal/observe/telemetry/metrics.go was a single 784-line var
block. Split into ten domain-scoped metrics_<area>.go files (request,
quota, rebalance, replication, breaker, cleanup, audit, encryption,
cache, meta) — each under ~155 lines. metrics.go retains only the
package doc.

Adds metrics_registration_test.go that touches a representative metric
from each domain file and asserts the family appears in
prometheus.DefaultGatherer.Gather() — guards against a domain file
being silently dropped or its var block deleted.
