+++ .version
@@ -1,1 +1,1 @@
-v0.42.2
+v0.42.3
