+++ internal/observe/telemetry/metrics.go
@@ -1,784 +1,14 @@
 // -------------------------------------------------------------------------------
-// Metrics - Prometheus Instrumentation
+// Metrics - Prometheus Instrumentation Index
 //
 // Author: Alex Freidah
 //
-// Prometheus metric definitions for the S3 orchestrator. Tracks request counts,
-// latencies, sizes, and backend health. All metrics are prefixed with 's3o_' for
-// easy identification in dashboards and alerting rules.
+// Subsystem-scoped Prometheus surface for the S3 orchestrator. The metric
+// definitions are split across metrics_<domain>.go files so each file stays
+// under ~150 lines and one subsystem's metrics live alongside each other.
+// All metrics are prefixed with "s3o_" and registered via promauto on init.
 // -------------------------------------------------------------------------------
 
 // Package telemetry provides Prometheus metrics registration and OpenTelemetry
 // tracing initialization for the S3 orchestrator.
 package telemetry
-
-import (
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
-)
-
-// -------------------------------------------------------------------------
-// METRIC DEFINITIONS
-// -------------------------------------------------------------------------
-
-var (
-	// --- Request metrics ---
-
-	// RequestsTotal counts all HTTP requests by method and status code.
-	RequestsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_requests_total",
-			Help: "Total number of HTTP requests processed",
-		},
-		[]string{"method", "status_code"},
-	)
-
-	// RequestDuration tracks request latency distribution by method.
-	RequestDuration = promauto.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "s3o_request_duration_seconds",
-			Help:    "HTTP request latency in seconds",
-			Buckets: []float64{.005, .01, .025, .05, .1, .25, .5, 1, 2.5, 5, 10, 30, 60},
-		},
-		[]string{"method"},
-	)
-
-	// RequestSize tracks upload sizes.
-	RequestSize = promauto.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "s3o_request_size_bytes",
-			Help:    "HTTP request body size in bytes",
-			Buckets: prometheus.ExponentialBuckets(1024, 4, 10), // 1KB to 256GB
-		},
-		[]string{"method"},
-	)
-
-	// ResponseSize tracks download sizes.
-	ResponseSize = promauto.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "s3o_response_size_bytes",
-			Help:    "HTTP response body size in bytes",
-			Buckets: prometheus.ExponentialBuckets(1024, 4, 10), // 1KB to 256GB
-		},
-		[]string{"method"},
-	)
-
-	// InflightRequests tracks currently processing requests.
-	InflightRequests = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_inflight_requests",
-			Help: "Number of requests currently being processed",
-		},
-		[]string{"method"},
-	)
-
-	// --- Backend metrics ---
-
-	// BackendRequestsTotal counts backend operations by operation type and status.
-	BackendRequestsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_backend_requests_total",
-			Help: "Total number of backend storage operations",
-		},
-		[]string{"operation", "backend", "status"},
-	)
-
-	// BackendDuration tracks backend operation latency.
-	BackendDuration = promauto.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "s3o_backend_duration_seconds",
-			Help:    "Backend operation latency in seconds",
-			Buckets: []float64{.01, .025, .05, .1, .25, .5, 1, 2.5, 5, 10, 30, 60, 120},
-		},
-		[]string{"operation", "backend"},
-	)
-
-	// --- Manager metrics ---
-
-	// ManagerRequestsTotal counts manager-level operations.
-	ManagerRequestsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_manager_requests_total",
-			Help: "Total number of manager-level storage operations",
-		},
-		[]string{"operation", "backend", "status"},
-	)
-
-	// ManagerDuration tracks manager operation latency.
-	ManagerDuration = promauto.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "s3o_manager_duration_seconds",
-			Help:    "Manager operation latency in seconds",
-			Buckets: []float64{.01, .025, .05, .1, .25, .5, 1, 2.5, 5, 10, 30, 60, 120},
-		},
-		[]string{"operation", "backend"},
-	)
-
-	// --- Quota metrics ---
-
-	// QuotaBytesUsed tracks current bytes used per backend.
-	QuotaBytesUsed = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_quota_bytes_used",
-			Help: "Current bytes used on each backend",
-		},
-		[]string{"backend"},
-	)
-
-	// QuotaBytesLimit tracks quota limit per backend.
-	QuotaBytesLimit = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_quota_bytes_limit",
-			Help: "Quota limit in bytes for each backend",
-		},
-		[]string{"backend"},
-	)
-
-	// QuotaBytesAvailable tracks available bytes per backend.
-	QuotaBytesAvailable = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_quota_bytes_available",
-			Help: "Available bytes (limit - used - orphan) for each backend",
-		},
-		[]string{"backend"},
-	)
-
-	// QuotaOrphanBytes tracks bytes pending physical deletion per backend.
-	QuotaOrphanBytes = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_quota_orphan_bytes",
-			Help: "Bytes pending physical deletion (logically freed but not yet removed from backend)",
-		},
-		[]string{"backend"},
-	)
-
-	// --- Object metrics ---
-
-	// ObjectCount tracks the number of objects stored per backend.
-	ObjectCount = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_objects_count",
-			Help: "Number of objects stored on each backend",
-		},
-		[]string{"backend"},
-	)
-
-	// --- Multipart metrics ---
-
-	// ActiveMultipartUploads tracks in-progress multipart uploads per backend.
-	ActiveMultipartUploads = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_active_multipart_uploads",
-			Help: "Number of in-progress multipart uploads per backend",
-		},
-		[]string{"backend"},
-	)
-
-	// --- Usage tracking metrics ---
-
-	// UsageAPIRequests tracks the current month's API request count per backend.
-	UsageAPIRequests = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_usage_api_requests",
-			Help: "Current month API request count per backend (from DB)",
-		},
-		[]string{"backend"},
-	)
-
-	// UsageEgressBytes tracks the current month's egress bytes per backend.
-	UsageEgressBytes = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_usage_egress_bytes",
-			Help: "Current month egress bytes per backend (from DB)",
-		},
-		[]string{"backend"},
-	)
-
-	// UsageIngressBytes tracks the current month's ingress bytes per backend.
-	UsageIngressBytes = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_usage_ingress_bytes",
-			Help: "Current month ingress bytes per backend (from DB)",
-		},
-		[]string{"backend"},
-	)
-
-	// --- Usage limit metrics ---
-
-	// UsageLimitRejectionsTotal counts operations rejected due to monthly usage limits.
-	UsageLimitRejectionsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_usage_limit_rejections_total",
-			Help: "Total operations rejected due to monthly usage limits",
-		},
-		[]string{"operation", "limit_type"},
-	)
-
-	// --- Rate limit metrics ---
-
-	// RateLimitRejectionsTotal counts requests rejected by the per-IP rate limiter.
-	RateLimitRejectionsTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_rate_limit_rejections_total",
-			Help: "Total requests rejected due to per-IP rate limiting",
-		},
-	)
-
-	// AdmissionRejectionsTotal counts requests rejected by server-level admission control.
-	AdmissionRejectionsTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_admission_rejections_total",
-			Help: "Total requests rejected due to server-level admission control",
-		},
-	)
-
-	// WorkerAdmissionRejectionsTotal counts background worker tasks that were
-	// skipped because the admission semaphore was full.
-	WorkerAdmissionRejectionsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_worker_admission_rejections_total",
-			Help: "Background worker tasks skipped due to admission control",
-		},
-		[]string{"worker"},
-	)
-
-	// LoadShedTotal counts requests probabilistically rejected by active
-	// load shedding before reaching the hard admission limit.
-	LoadShedTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_load_shed_total",
-			Help: "Requests shed by active load shedding before hard admission limit",
-		},
-	)
-
-	// EarlyRejectionsTotal counts uploads rejected before body transmission
-	// via Expect: 100-Continue pre-flight capacity checks.
-	EarlyRejectionsTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_early_rejections_total",
-			Help: "Uploads rejected before body transmission (no backend capacity)",
-		},
-	)
-
-	// --- Rebalancer metrics ---
-
-	// RebalanceObjectsMoved counts objects moved by the rebalancer.
-	RebalanceObjectsMoved = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_rebalance_objects_moved_total",
-			Help: "Total number of objects moved by the rebalancer",
-		},
-		[]string{"strategy", "status"},
-	)
-
-	// RebalanceBytesMoved counts bytes moved by the rebalancer.
-	RebalanceBytesMoved = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_rebalance_bytes_moved_total",
-			Help: "Total bytes moved by the rebalancer",
-		},
-		[]string{"strategy"},
-	)
-
-	// RebalanceRunsTotal counts rebalancer executions.
-	RebalanceRunsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_rebalance_runs_total",
-			Help: "Total number of rebalancer runs",
-		},
-		[]string{"strategy", "status"},
-	)
-
-	// RebalanceDuration tracks rebalancer execution time.
-	RebalanceDuration = promauto.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "s3o_rebalance_duration_seconds",
-			Help:    "Rebalancer execution time in seconds",
-			Buckets: []float64{1, 5, 10, 30, 60, 120, 300, 600},
-		},
-		[]string{"strategy"},
-	)
-
-	// RebalanceSkipped counts rebalancer runs that were skipped.
-	RebalanceSkipped = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_rebalance_skipped_total",
-			Help: "Total number of rebalancer runs skipped",
-		},
-		[]string{"reason"},
-	)
-
-	// RebalancePending tracks objects planned for rebalance in the current cycle.
-	RebalancePending = promauto.NewGauge(
-		prometheus.GaugeOpts{
-			Name: "s3o_rebalance_pending",
-			Help: "Number of objects planned for rebalance",
-		},
-	)
-
-	// --- Replication metrics ---
-
-	// ReplicationPending tracks objects currently below the target replication factor.
-	ReplicationPending = promauto.NewGauge(
-		prometheus.GaugeOpts{
-			Name: "s3o_replication_pending",
-			Help: "Number of objects below the target replication factor",
-		},
-	)
-
-	// ReplicationCopiesCreatedTotal counts replica copies created.
-	ReplicationCopiesCreatedTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_replication_copies_created_total",
-			Help: "Total number of replica copies created",
-		},
-	)
-
-	// ReplicationErrorsTotal counts replication errors.
-	ReplicationErrorsTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_replication_errors_total",
-			Help: "Total number of replication errors",
-		},
-	)
-
-	// ReplicationDuration tracks replication worker cycle time.
-	ReplicationDuration = promauto.NewHistogram(
-		prometheus.HistogramOpts{
-			Name:    "s3o_replication_duration_seconds",
-			Help:    "Replication worker cycle time in seconds",
-			Buckets: []float64{1, 5, 10, 30, 60, 120, 300, 600},
-		},
-	)
-
-	// ReplicationRunsTotal counts replication worker executions.
-	ReplicationRunsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_replication_runs_total",
-			Help: "Total number of replication worker runs",
-		},
-		[]string{"status"},
-	)
-
-	// ReplicationHealthCopiesTotal counts copies created to replace copies on
-	// circuit-broken backends.
-	ReplicationHealthCopiesTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_replication_health_copies_total",
-			Help: "Replica copies created to replace copies on circuit-broken backends",
-		},
-	)
-
-	// --- Over-replication cleanup metrics ---
-
-	// OverReplicationPending tracks objects currently above the target replication factor.
-	OverReplicationPending = promauto.NewGauge(
-		prometheus.GaugeOpts{
-			Name: "s3o_over_replication_pending",
-			Help: "Number of objects above the target replication factor",
-		},
-	)
-
-	// OverReplicationRemovedTotal counts excess copies removed by the cleaner.
-	OverReplicationRemovedTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_over_replication_removed_total",
-			Help: "Total excess copies removed by over-replication cleanup",
-		},
-	)
-
-	// OverReplicationErrorsTotal counts over-replication cleanup errors.
-	OverReplicationErrorsTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_over_replication_errors_total",
-			Help: "Total number of over-replication cleanup errors",
-		},
-	)
-
-	// OverReplicationRunsTotal counts over-replication cleanup worker executions.
-	OverReplicationRunsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_over_replication_runs_total",
-			Help: "Total number of over-replication cleanup worker runs",
-		},
-		[]string{"status"},
-	)
-
-	// OverReplicationDuration tracks over-replication cleanup worker cycle time.
-	OverReplicationDuration = promauto.NewHistogram(
-		prometheus.HistogramOpts{
-			Name:    "s3o_over_replication_duration_seconds",
-			Help:    "Over-replication cleanup worker cycle time in seconds",
-			Buckets: []float64{1, 5, 10, 30, 60, 120, 300, 600},
-		},
-	)
-
-	// --- Circuit breaker metrics ---
-
-	// CircuitBreakerState tracks the current circuit breaker state per component.
-	// 0=closed (healthy), 1=open (down), 2=half-open (probing).
-	CircuitBreakerState = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_circuit_breaker_state",
-			Help: "Current circuit breaker state: 0=closed, 1=open, 2=half-open",
-		},
-		[]string{"name"},
-	)
-
-	// CircuitBreakerTransitionsTotal counts state transitions per component.
-	CircuitBreakerTransitionsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_circuit_breaker_transitions_total",
-			Help: "Total number of circuit breaker state transitions",
-		},
-		[]string{"name", "from", "to"},
-	)
-
-	// CircuitBreakerInternalErrorsTotal counts errors returned by the
-	// breaker's own machinery (PostCheck / state transition helpers).
-	// Non-zero values indicate a bookkeeping bug, not an application
-	// error — alert on any increase.
-	CircuitBreakerInternalErrorsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_circuit_breaker_internal_errors_total",
-			Help: "Errors returned by circuit breaker PostCheck/state transitions",
-		},
-		[]string{"name", "operation"},
-	)
-
-	// DegradedReadsTotal counts reads served via broadcast during degraded mode.
-	DegradedReadsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_degraded_reads_total",
-			Help: "Total number of read operations served via broadcast during degraded mode",
-		},
-		[]string{"operation"},
-	)
-
-	// DegradedCacheHitsTotal counts location cache hits during degraded reads.
-	DegradedCacheHitsTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_degraded_cache_hits_total",
-			Help: "Total number of location cache hits during degraded reads",
-		},
-	)
-
-	// DegradedWriteRejectionsTotal counts writes rejected during degraded mode.
-	DegradedWriteRejectionsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_degraded_write_rejections_total",
-			Help: "Total number of write operations rejected during degraded mode",
-		},
-		[]string{"operation"},
-	)
-
-	// WriteFailoverTotal counts writes that failed on one backend and were
-	// retried on another. Labels: operation, failed_backend, success_backend.
-	WriteFailoverTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_write_failover_total",
-			Help: "Total number of write operations that failed over to a different backend",
-		},
-		[]string{"operation", "failed_backend", "success_backend"},
-	)
-
-	// --- Cleanup queue metrics ---
-
-	// CleanupQueueEnqueuedTotal counts items added to the cleanup retry queue.
-	CleanupQueueEnqueuedTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_cleanup_queue_enqueued_total",
-			Help: "Total items added to the cleanup retry queue",
-		},
-		[]string{"reason"},
-	)
-
-	// CleanupQueueProcessedTotal counts items processed from the cleanup queue.
-	CleanupQueueProcessedTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_cleanup_queue_processed_total",
-			Help: "Total items processed from the cleanup retry queue",
-		},
-		[]string{"status"},
-	)
-
-	// CleanupQueueDepth tracks the current number of pending cleanup items.
-	CleanupQueueDepth = promauto.NewGauge(
-		prometheus.GaugeOpts{
-			Name: "s3o_cleanup_queue_depth",
-			Help: "Current number of pending items in the cleanup retry queue",
-		},
-	)
-
-	// --- Lifecycle metrics ---
-
-	// LifecycleDeletedTotal counts objects deleted by lifecycle expiration rules.
-	LifecycleDeletedTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_lifecycle_deleted_total",
-			Help: "Objects deleted by lifecycle expiration rules",
-		},
-	)
-
-	// LifecycleFailedTotal counts objects that failed lifecycle deletion.
-	LifecycleFailedTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_lifecycle_failed_total",
-			Help: "Objects that failed lifecycle deletion",
-		},
-	)
-
-	// LifecycleRunsTotal counts lifecycle worker executions.
-	LifecycleRunsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_lifecycle_runs_total",
-			Help: "Lifecycle worker executions",
-		},
-		[]string{"status"},
-	)
-
-	// --- Audit metrics ---
-
-	// AuditEventsTotal counts audit log entries by event type.
-	AuditEventsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_audit_events_total",
-			Help: "Total number of audit log entries emitted",
-		},
-		[]string{"event"},
-	)
-
-	// --- Drain metrics ---
-
-	// DrainObjectsMoved counts objects moved during backend drain operations.
-	DrainObjectsMoved = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_drain_objects_moved_total",
-			Help: "Total number of objects moved during backend drain operations",
-		},
-	)
-
-	// DrainBytesMoved counts bytes moved during backend drain operations.
-	DrainBytesMoved = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_drain_bytes_moved_total",
-			Help: "Total bytes moved during backend drain operations",
-		},
-	)
-
-	// DrainActive is 1 when a drain operation is in progress, 0 otherwise.
-	DrainActive = promauto.NewGauge(
-		prometheus.GaugeOpts{
-			Name: "s3o_drain_active",
-			Help: "Whether a backend drain operation is currently in progress",
-		},
-	)
-
-	// --- Encryption metrics ---
-
-	// EncryptionOpsTotal counts encryption operations by type.
-	EncryptionOpsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_encryption_operations_total",
-			Help: "Total encryption operations",
-		},
-		[]string{"op"},
-	)
-
-	// EncryptionErrorsTotal counts encryption errors by operation and type.
-	EncryptionErrorsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_encryption_errors_total",
-			Help: "Total encryption errors",
-		},
-		[]string{"op", "error_type"},
-	)
-
-	// EncryptionUnknownKeyIDTotal counts decryption attempts where the keyID
-	// was not found in the configured keys, triggering a primary key fallback.
-	EncryptionUnknownKeyIDTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_encryption_unknown_key_id_total",
-			Help: "Decryption attempts with unknown keyID (primary key fallback)",
-		},
-	)
-
-	// --- Integrity verification metrics ---
-
-	// IntegrityErrorsTotal counts hash mismatches detected during read,
-	// replication, or background scrubbing.
-	IntegrityErrorsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_integrity_errors_total",
-			Help: "Content hash mismatches detected",
-		},
-		[]string{"operation"},
-	)
-
-	// IntegrityChecksTotal counts hash verifications performed.
-	IntegrityChecksTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_integrity_checks_total",
-			Help: "Content hash verifications performed",
-		},
-		[]string{"operation"},
-	)
-
-	// KeyRotationObjectsTotal counts objects processed during key rotation.
-	KeyRotationObjectsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_key_rotation_objects_total",
-			Help: "Total objects processed during key rotation",
-		},
-		[]string{"status"},
-	)
-
-	// EncryptExistingObjectsTotal counts objects processed during encrypt-existing.
-	EncryptExistingObjectsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_encrypt_existing_objects_total",
-			Help: "Total objects processed during encrypt-existing operation",
-		},
-		[]string{"status"},
-	)
-
-	// DecryptExistingObjectsTotal counts objects processed during decrypt-existing.
-	DecryptExistingObjectsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_decrypt_existing_objects_total",
-			Help: "Total objects processed during decrypt-existing operation",
-		},
-		[]string{"status"},
-	)
-
-	// --- Cache metrics ---
-
-	// CacheHitsTotal counts object data cache hits.
-	CacheHitsTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_cache_hits_total",
-			Help: "Object data cache hits",
-		},
-	)
-
-	// CacheMissesTotal counts object data cache misses.
-	CacheMissesTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_cache_misses_total",
-			Help: "Object data cache misses",
-		},
-	)
-
-	// CacheEvictionsTotal counts cache entries evicted by LRU or TTL.
-	CacheEvictionsTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_cache_evictions_total",
-			Help: "Cache entries evicted (LRU or TTL)",
-		},
-	)
-
-	// CacheSizeBytes tracks current cache utilization in bytes.
-	CacheSizeBytes = promauto.NewGauge(
-		prometheus.GaugeOpts{
-			Name: "s3o_cache_size_bytes",
-			Help: "Current object data cache size in bytes",
-		},
-	)
-
-	// CacheEntries tracks the number of entries in the cache.
-	CacheEntries = promauto.NewGauge(
-		prometheus.GaugeOpts{
-			Name: "s3o_cache_entries",
-			Help: "Number of entries in the object data cache",
-		},
-	)
-
-	// --- Redis metrics ---
-
-	// RedisOperationsTotal counts Redis counter backend operations.
-	RedisOperationsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_redis_operations_total",
-			Help: "Total Redis counter backend operations",
-		},
-		[]string{"operation", "status"},
-	)
-
-	// RedisFallbackActive is 1 when the Redis counter backend is in local
-	// fallback mode due to circuit breaker, 0 during normal operation.
-	RedisFallbackActive = promauto.NewGauge(
-		prometheus.GaugeOpts{
-			Name: "s3o_redis_fallback_active",
-			Help: "Whether Redis counter backend is in local fallback mode",
-		},
-	)
-
-	// --- Info metric ---
-
-	// BuildInfo exposes version information.
-	BuildInfo = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "s3o_build_info",
-			Help: "Build information for the S3 proxy",
-		},
-		[]string{"version", "go_version"},
-	)
-
-	// --- Notification metrics ---
-
-	// NotificationSentTotal counts successfully delivered webhook notifications.
-	NotificationSentTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_notification_sent_total",
-			Help: "Webhook notifications delivered successfully",
-		},
-		[]string{"endpoint", "event_type"},
-	)
-
-	// NotificationFailedTotal counts webhook delivery failures.
-	NotificationFailedTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_notification_failed_total",
-			Help: "Webhook notification delivery failures",
-		},
-		[]string{"endpoint", "event_type"},
-	)
-
-	// NotificationDroppedTotal counts events dropped due to dampening or enqueue failure.
-	NotificationDroppedTotal = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Name: "s3o_notification_dropped_total",
-			Help: "Events dropped due to dampening or queue insertion failure",
-		},
-	)
-
-	// NotificationStoreErrorsTotal counts outbox-store operation failures in
-	// the delivery worker (CompleteNotification / RetryNotification). A
-	// non-zero value means the worker saw a store error that could cause
-	// duplicate or dropped webhook deliveries — alert on any increase.
-	NotificationStoreErrorsTotal = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "s3o_notification_store_errors_total",
-			Help: "Outbox store errors seen by the notification delivery worker",
-		},
-		[]string{"operation"},
-	)
-
-	// NotificationQueueDepth reports the number of pending notifications in the outbox.
-	NotificationQueueDepth = promauto.NewGauge(
-		prometheus.GaugeOpts{
-			Name: "s3o_notification_queue_depth",
-			Help: "Pending notifications in the delivery outbox",
-		},
-	)
-
-	// NotificationDuration measures webhook delivery latency.
-	NotificationDuration = promauto.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "s3o_notification_duration_seconds",
-			Help:    "Webhook notification delivery latency",
-			Buckets: prometheus.DefBuckets,
-		},
-		[]string{"endpoint"},
-	)
-)
